- refactor project structure
