+++ api/src/handlers/process.rs
@@ -1,0 +1,77 @@
+use crate::handlers::types::{ApiCommand, ApiCommandResult};
+use crate::worker::{ProcessState, WorkerEngine};
+use rocket::State;
+use uuid::Uuid;
+
+#[get("/process_status/<process_id>")]
+pub async fn get_process_status(process_id: String, engine: &State<WorkerEngine>) -> String {
+    // get status of process by ID
+    match Uuid::parse_str(&process_id) {
+        Ok(process_uuid) => {
+            if engine.arc_process_states.contains_key(&process_uuid) {
+                format!(
+                    "{:}",
+                    engine
+                        .arc_process_states
+                        .get(&process_uuid)
+                        .unwrap()
+                        .value()
+                )
+            } else {
+                // TODO can we return HTTP status code here?
+                format!("Process id not found")
+            }
+        }
+        Err(e) => {
+            // TODO can we return HTTP status code here?
+            e.to_string()
+        }
+    }
+}
+
+pub fn do_process_command(command: ApiCommand, engine: &State<WorkerEngine>) -> String {
+    // queue the new Scarb command
+    match engine.enqueue_command(command) {
+        Ok(uuid) => {
+            // return the process ID
+            format!("{}", uuid)
+        }
+        Err(e) => {
+            // TODO can we return HTTP status code here?
+            e
+        }
+    }
+}
+
+pub fn fetch_process_result<F>(
+    process_id: String,
+    engine: &State<WorkerEngine>,
+    do_work: F,
+) -> String
+where
+    F: FnOnce(&ApiCommandResult) -> String,
+{
+    // get status of process by ID
+    match Uuid::parse_str(&process_id) {
+        Ok(process_uuid) => {
+            if engine.arc_process_states.contains_key(&process_uuid) {
+                match engine
+                    .arc_process_states
+                    .get(&process_uuid)
+                    .unwrap()
+                    .value()
+                {
+                    ProcessState::Completed(result) => do_work(result),
+                    _ => String::from("Result not available"),
+                }
+            } else {
+                // TODO can we return HTTP status code here?
+                "Process id not found".to_string()
+            }
+        }
+        Err(e) => {
+            // TODO can we return HTTP status code here?
+            e.to_string()
+        }
+    }
+}
