+++ api/src/handlers/types.rs
@@ -1,0 +1,42 @@
+use serde::{Deserialize, Serialize};
+use std::path::PathBuf;
+
+#[derive(Debug, Deserialize, Serialize)]
+#[serde(crate = "rocket::serde")]
+pub struct CompileResponse {
+    pub status: String,
+    pub message: String,
+    pub file_content: String,
+}
+
+#[derive(Serialize, Deserialize)]
+pub struct FileContentMap {
+    pub file_name: String,
+    pub file_content: String,
+}
+
+#[derive(Serialize, Deserialize)]
+pub struct ScarbCompileResponse {
+    pub status: String,
+    pub message: String,
+    pub file_content_map_array: Vec<FileContentMap>,
+}
+
+#[derive(Debug)]
+pub enum ApiCommand {
+    CairoVersion,
+    SierraCompile(PathBuf),
+    CasmCompile(PathBuf),
+    ScarbCompile(PathBuf),
+    #[allow(dead_code)]
+    Shutdown,
+}
+
+pub enum ApiCommandResult {
+    CairoVersion(String),
+    CasmCompile(CompileResponse),
+    SierraCompile(CompileResponse),
+    ScarbCompile(ScarbCompileResponse),
+    #[allow(dead_code)]
+    Shutdown,
+}
